requirements.txt
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
joblib==1.2.0
pyserial==3.5